src/main/java/lotto/model/LottoTicket.java
package lotto.model;

import java.util.*;

public class LottoTicket {

    private List<Lotto> lottos;

    public LottoTicket(List<Lotto> lottos) {
        this.lottos = lottos;
    }

    public static LottoTicket of(List<Lotto> lottos) {
        return new LottoTicket(lottos);
    }


    public List<Lotto> getLottos() {
        return Collections.unmodifiableList(lottos);
    }

    public LottoResult result(Lotto winningLotto) {
        List<Prize> prizes = new ArrayList<>();
        for (Lotto lotto : lottos) {
            prizes.add(Prize.of(lotto.getMatchCount(winningLotto)));
        }
        return LottoResult.of(prizes);
    }
}
